requirements_cpu.txt
absl-py==0.10.0
alembic==1.5.8
astunparse==1.6.3
Brotli==1.0.9
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.6
cycler==0.10.0
dash==1.19.0
dash-bootstrap-components==0.11.3
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dnspython==2.1.0
easydict==1.9
email-validator==1.1.2
Flask==1.1.2
Flask-Assets==2.0
Flask-Compress==1.9.0
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.2
Mako==1.1.4
Markdown==3.3
MarkupSafe==1.1.1
matplotlib @ file:///C:/ci/matplotlib-base_1597858387645/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1596233945180/work
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.1.26
opt-einsum==3.3.0
pandas==1.1.5
Pillow @ file:///C:/ci/pillow_1594298234712/work
plotly==4.14.3
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0rc0
termcolor==1.1.0
tf-estimator-nightly==2.3.0.dev2020062301
tornado==6.0.4
tqdm==4.50.2
urllib3==1.25.10
webassets==2.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
zipp==3.3.0
